feat: Improve notification badge display and update event notifications label
This commit improves the display of the notification badge by dynamically showing the number of event notifications. If there are more than 9 notifications, the badge content will be displayed as '9+'. Additionally, the label for the event notifications in the notification dropdown has been updated to show the actual number of new notifications. These changes enhance the user experience by providing more accurate and informative notifications.🔔📝📢